Add function that calculates box placement
The calculation is made to offset the box slightly to the right of the
cursor, with the first option being on the same level as it.

This changes if the box would overflow to the bottom or to the right, in
which case values are recalcuated to place the last value at the same
level as the cursor, to the left of the cursor, or both.