Add a controller to track the time
A request is made from the front end to set a start time stored within
the session cache as the current minutes and seconds.

I have chosen the session cache as the storage because this value is only
needed once per game and does not need to be stored permanently in the
database.