Add the ability for users to store their scores
The ScoreForm react component, LeaderboardPost model and LeaderboardPosts
controller were created for this reason.

The ScoreForm gets displayed within the Image component once the win
condition is met and on submit, it makes a post request to the create
method of the LeaderboardPosts controller. It still needs frontend
validation, but that will be done on the next commit.