Add functioning top 10 leaderboard
To do this I have changed completion_time into a method that return the
time in seconds for easier sortin.

I am looking for the first 10 scores in the top_ten method
from the LeaderboardPost controller and fetching them in my Leaderboard
component on the frontEnd.

I have moved the formatting of the completion_time
into its separate function on the frontend, since I only need the formatted
time for display purposes.